BAEL-3482
Code re-factor
